feat: add a make_admin binary to help with startup
Summary:
- This was listed as an issue for the last startup.
- This PR moves the functionality from pruely MIX based to a command
  line tool tool.

Test Plan:
Used it locally
